file strcuture refactor| removing unsed files
