+++ App.tsx
@@ -1,46 +1,20 @@
 import React, { useEffect, useState } from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import Home from "./src/screens/Home";
-import PostDetails from "./src/screens/PostDetails";
-import NotFound from "./src/screens/NotFound";
-import LoginScreen from "./src/screens/Login";
 import {
-  RootStackParamList,
   SetupStackParamList,
-  HomeTabParamList,
 } from "./src/navigation/types";
-import SplashScreen from "./src/screens/Splash";
-import RegisterScreen from "./src/screens/Register";
-import ForgotScreen from "./src/screens/Forgot";
 import { ThemeProvider } from "./src/state/theme";
 import { NativeWindStyleSheet } from "nativewind";
+import SplashScreen from "./src/screens/setup/Splash";
+import LoginScreen from "./src/screens/setup/Login";
+import RegisterScreen from "./src/screens/setup/Register";
+import ForgotScreen from "./src/screens/setup/Forgot";
 
 NativeWindStyleSheet.setOutput({default: "native"})
 
-const Stack = createStackNavigator<RootStackParamList>();
 const SetupStack = createStackNavigator<SetupStackParamList>();
-const Tab = createBottomTabNavigator<HomeTabParamList>();
 
-function HomeStack() {
-  return (
-    <Stack.Navigator initialRouteName="Home">
-      <Stack.Screen name="Home" component={Home} />
-      <Stack.Screen name="PostDetails" component={PostDetails} />
-      <Stack.Screen name="NotFound" component={NotFound} />
-    </Stack.Navigator>
-  );
-}
-
-function RootNavigator() {
-  return (
-    <Tab.Navigator>
-      <Tab.Screen name="Popular" component={HomeStack} />
-      <Tab.Screen name="Latest" component={PostDetails} />
-    </Tab.Navigator>
-  );
-}
 
 export default function App() {
   const [isAuthenticated, setIsAuthenticated] = useState(false);
@@ -85,10 +59,4 @@ export default function App() {
       </ThemeProvider>
     );
   }
-
-  return (
-    <NavigationContainer>
-      <RootNavigator />
-    </NavigationContainer>
-  );
 }
