+++ src/navigation/types.tsx
@@ -1,15 +1,8 @@
-import type {
-  CompositeScreenProps,
-  NavigatorScreenParams,
-} from "@react-navigation/native";
 import type { StackScreenProps } from "@react-navigation/stack";
-import type { BottomTabScreenProps } from "@react-navigation/bottom-tabs";
 
 declare global {
   namespace ReactNavigation {
-    interface RootParamList extends RootStackParamList {}
     interface SetupParamList extends SetupStackParamList {}
-    interface HomeParamList extends HomeTabParamList {}
   }
 }
 
@@ -23,22 +16,3 @@ export type SetupStackParamList = {
 export type SetupStackScreenProps<T extends keyof SetupStackParamList> =
   StackScreenProps<SetupStackParamList, T>;
 
-export type RootStackParamList = {
-  Home: NavigatorScreenParams<HomeTabParamList>;
-  PostDetails: { id: string };
-  NotFound: undefined;
-};
-
-export type RootStackScreenProps<T extends keyof RootStackParamList> =
-  StackScreenProps<RootStackParamList, T>;
-
-export type HomeTabParamList = {
-  Popular: undefined;
-  Latest: undefined;
-};
-
-export type HomeTabScreenProps<T extends keyof HomeTabParamList> =
-  CompositeScreenProps<
-    BottomTabScreenProps<HomeTabParamList, T>,
-    RootStackScreenProps<keyof RootStackParamList>
-  >;
