--- src/screens/Home.tsx
@@ -1,10 +1,0 @@
-import { SafeAreaView, Text } from "react-native";
-
-
-export default function Home() {
-    return (
-        <SafeAreaView>
-            <Text>My home screen</Text>
-        </SafeAreaView>
-    )
-}
