--- src/screens/PostDetails.tsx
@@ -1,10 +1,0 @@
-import { SafeAreaView, Text } from "react-native";
-
-export default function PostDetails() {
-     return (
-        <SafeAreaView>
-            <Text>My Post screen</Text>
-        </SafeAreaView>
-    )
-
-}
