+++ src/screens/setup/Register.tsx
@@ -5,9 +5,8 @@ import {
   Text,
   TextInput,
   TouchableOpacity,
-  StyleSheet,
 } from "react-native";
-import { useTheme } from "../state/theme";
+import { useTheme } from "../../state/theme";
 
 export default function RegisterScreen({ navigation }) {
   const [name, setName] = useState("");
@@ -25,67 +24,11 @@ export default function RegisterScreen({ navigation }) {
     console.log("User Type:", userType);
   };
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: darkMode ? "#black" : "F6F6F6",
-    paddingHorizontal: 20,
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: "bold",
-    marginBottom: 20,
-  },
-  input: {
-    width: "100%",
-    height: 40,
-    borderWidth: 1,
-    borderColor: "#E4E4E4",
-    borderRadius: 5,
-    marginBottom: 10,
-    paddingLeft: 10,
-    color: "#333333",
-    backgroundColor: "#FFFFFF",
-  },
-  pickerContainer: {
-    width: "100%",
-    marginBottom: 20,
-  },
-  label: {
-    fontSize: 16,
-    marginBottom: 10,
-  },
-  picker: {
-    width: "100%",
-    height: 30,
-    borderWidth: 1,
-    borderColor: "#E4E4E4",
-    borderRadius: 5,
-    backgroundColor: "#FFFFFF",
-    marginBottom: 20,
-  },
-  registerButton: {
-    backgroundColor: "#4285F4",
-    paddingVertical: 10,
-    width: "100%",
-    borderRadius: 5,
-  },
-  registerButtonText: {
-    color: "#FFFFFF",
-    fontSize: 16,
-    textAlign: "center",
-  },
-});
-
-
   return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Cadastro</Text>
+    <View className={`flex justify-center items-center h-full ${darkMode ? 'bg-gray-900' : 'bg-gray-100'}`}>
+      <Text >Cadastro</Text>
 
       <TextInput
-        style={styles.input}
         placeholder="Nome"
         onChangeText={(text) => setName(text)}
         value={name}
@@ -93,7 +36,6 @@ const styles = StyleSheet.create({
       />
 
       <TextInput
-        style={styles.input}
         placeholder="Email"
         onChangeText={(text) => setEmail(text)}
         value={email}
@@ -102,7 +44,6 @@ const styles = StyleSheet.create({
       />
 
       <TextInput
-        style={styles.input}
         placeholder="Senha"
         onChangeText={(text) => setPassword(text)}
         value={password}
@@ -110,10 +51,9 @@ const styles = StyleSheet.create({
         placeholderTextColor="#B3B3B3"
       />
 
-      <View style={styles.pickerContainer}>
+      <View className="flex w-5/6" >
         <Picker
           selectedValue={userType}
-          style={styles.picker}
           onValueChange={(itemValue) => setUserType(itemValue)}
         >
           <Picker.Item label="Tosador" value="tosador" />
@@ -121,8 +61,8 @@ const styles = StyleSheet.create({
         </Picker>
       </View>
 
-      <TouchableOpacity style={styles.registerButton} onPress={handleRegister}>
-        <Text style={styles.registerButtonText}>Registrar</Text>
+      <TouchableOpacity  onPress={handleRegister}>
+        <Text >Registrar</Text>
       </TouchableOpacity>
     </View>
   );
